+++ d2.config.js
@@ -1,10 +1,11 @@
 const config = {
     type: 'app',
-    coreApp: true,
+    name: 'dashboard',
     title: 'Dashboard',
+    coreApp: true,
 
     entryPoints: {
-        app: './src/AppWrapper',
+        app: './src/AppWrapper.js',
     },
 }