+++ src/components/Item/VisualizationItem/ItemContextMenu/ViewAsMenuItems.js
@@ -4,6 +4,7 @@ import i18n from '@dhis2/d2-i18n'
 import TableIcon from '@material-ui/icons/ViewList'
 import ChartIcon from '@material-ui/icons/InsertChart'
 import MapIcon from '@material-ui/icons/Public'
+import getThematicMapViews from '../../../../modules/getThematicMapViews'
 import {
     CHART,
     MAP,
@@ -13,11 +14,16 @@ import {
     isTrackerDomainType,
     hasMapView,
 } from '../../../../modules/itemTypes'
-import { MenuItem, colors } from '@dhis2/ui'
+import { MenuItem, Tooltip, colors } from '@dhis2/ui'
 
 const iconFill = { fill: colors.grey600 }
 
-const ViewAsMenuItems = ({ type, activeType, onActiveTypeChanged }) => {
+const ViewAsMenuItems = ({
+    type,
+    activeType,
+    visualization,
+    onActiveTypeChanged,
+}) => {
     const isTrackerType = isTrackerDomainType(type)
 
     const onViewTable = () =>
@@ -28,23 +34,63 @@ const ViewAsMenuItems = ({ type, activeType, onActiveTypeChanged }) => {
 
     const onViewMap = () => onActiveTypeChanged(MAP)
 
+    const isDisabled = type === MAP && !getThematicMapViews(visualization)
+
+    const ViewAsChartMenuItem = () => {
+        const ChartMenuItem = () => (
+            <MenuItem
+                dense
+                disabled={isDisabled}
+                label={i18n.t('View as Chart')}
+                onClick={onViewChart}
+                icon={<ChartIcon style={iconFill} />}
+            />
+        )
+
+        if (isDisabled) {
+            return (
+                <Tooltip
+                    content={i18n.t("This map can't be displayed as a chart")}
+                >
+                    <ChartMenuItem />
+                </Tooltip>
+            )
+        }
+
+        return <ChartMenuItem />
+    }
+
+    const ViewAsTableMenuItem = () => {
+        const TableMenuItem = () => (
+            <MenuItem
+                dense
+                disabled={isDisabled}
+                label={i18n.t('View as Table')}
+                onClick={onViewTable}
+                icon={<TableIcon style={iconFill} />}
+            />
+        )
+
+        if (isDisabled) {
+            return (
+                <Tooltip
+                    content={i18n.t("This map can't be displayed as a table")}
+                >
+                    <TableMenuItem />
+                </Tooltip>
+            )
+        }
+
+        return <TableMenuItem />
+    }
+
     return (
         <>
             {activeType !== CHART && activeType !== EVENT_CHART && (
-                <MenuItem
-                    dense
-                    label={i18n.t('View as Chart')}
-                    onClick={onViewChart}
-                    icon={<ChartIcon style={iconFill} />}
-                />
+                <ViewAsChartMenuItem />
             )}
             {activeType !== REPORT_TABLE && activeType !== EVENT_REPORT && (
-                <MenuItem
-                    dense
-                    label={i18n.t('View as Table')}
-                    onClick={onViewTable}
-                    icon={<TableIcon style={iconFill} />}
-                />
+                <ViewAsTableMenuItem />
             )}
             {hasMapView(type) && activeType !== MAP && (
                 <MenuItem
@@ -61,6 +107,7 @@ const ViewAsMenuItems = ({ type, activeType, onActiveTypeChanged }) => {
 ViewAsMenuItems.propTypes = {
     activeType: PropTypes.string,
     type: PropTypes.string,
+    visualization: PropTypes.object,
     onActiveTypeChanged: PropTypes.func,
 }