+++ src/components/Item/VisualizationItem/ItemContextMenu/__tests__/ItemContextMenu.spec.js
@@ -1,9 +1,10 @@
 import React from 'react'
-import { render, waitFor } from '@testing-library/react'
+import { render, waitFor, screen } from '@testing-library/react'
 import { fireEvent } from '@testing-library/dom'
 import WindowDimensionsProvider from '../../../../WindowDimensionsProvider'
 import { useSystemSettings } from '../../../../SystemSettingsProvider'
 import ItemContextMenu from '../ItemContextMenu'
+import { getGridItemDomElementClassName } from '../../../../../modules/getGridItemDomElementClassName'
 
 jest.mock('../../../../SystemSettingsProvider', () => ({
     useSystemSettings: jest.fn(),
@@ -21,6 +22,7 @@ const mockSystemSettingsDefault = {
 const defaultProps = {
     item: {
         type: 'CHART',
+        id: 'rainbowdash',
     },
     visualization: {
         type: 'BAR',
@@ -30,7 +32,7 @@ const defaultProps = {
     activeFooter: false,
     activeType: 'CHART',
     fullscreenSupported: true,
-    isFullscreen: false,
+    loadItemFailed: false,
 }
 
 test('renders just the button when menu closed', () => {
@@ -46,17 +48,31 @@ test('renders just the button when menu closed', () => {
 })
 
 test('renders exit fullscreen button', () => {
-    useSystemSettings.mockImplementationOnce(() => mockSystemSettingsDefault)
-    const props = Object.assign({}, defaultProps, {
-        isFullscreen: true,
-    })
-    const { container } = render(
+    useSystemSettings.mockImplementation(() => mockSystemSettingsDefault)
+    const gridItemClassName = getGridItemDomElementClassName(
+        defaultProps.item.id
+    )
+
+    const { rerender } = render(
         <WindowDimensionsProvider>
-            <ItemContextMenu {...props} />
+            <div className={gridItemClassName}>
+                <ItemContextMenu {...defaultProps} />
+            </div>
         </WindowDimensionsProvider>
     )
 
-    expect(container).toMatchSnapshot()
+    document.fullscreenElement = document.querySelector(`.${gridItemClassName}`)
+
+    rerender(
+        <WindowDimensionsProvider>
+            <div className={{ gridItemClassName }}>
+                <ItemContextMenu {...defaultProps} />
+            </div>
+        </WindowDimensionsProvider>
+    )
+
+    document.fullscreenElement = null
+    expect(screen.getByTestId('exit-fullscreen-button')).toBeTruthy()
 })
 
 test('renders popover menu for BAR chart', async () => {
@@ -311,6 +327,36 @@ test('does not render "Open in [app]" option if settings do not allow', async ()
 
     await waitFor(() => {
         expect(queryByText(/Open in/i)).toBeNull()
+    })
+})
+
+test('renders only View in App when item load failed', async () => {
+    useSystemSettings.mockImplementation(() => mockSystemSettingsDefault)
+    const props = Object.assign({}, defaultProps, {
+        item: {
+            type: 'MAP',
+        },
+        visualization: {},
+        activeType: 'MAP',
+        loadItemFailed: true,
+    })
+
+    const { getByRole, queryByTestId, queryByText } = render(
+        <WindowDimensionsProvider>
+            <ItemContextMenu {...props} />
+        </WindowDimensionsProvider>
+    )
+
+    fireEvent.click(getByRole('button'))
+
+    await waitFor(() => {
+        expect(queryByText('View as Map')).toBeNull()
+        expect(queryByText('View as Chart')).toBeNull()
+        expect(queryByText('View as Table')).toBeNull()
+        expect(queryByTestId('divider')).toBeNull()
+        expect(queryByText('Open in Maps app')).toBeTruthy()
+        expect(queryByText('Show interpretations and details')).toBeNull()
+        expect(queryByText('View fullscreen')).toBeNull()
     })
 })