+++ src/components/Item/VisualizationItem/Visualization/LoadingMask.js
@@ -1,14 +1,19 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import { CircularLoader } from '@dhis2/ui'
 
 import classes from './styles/LoadingMask.module.css'
 
-const LoadingMask = () => {
+const LoadingMask = ({ style }) => {
     return (
-        <div className={classes.center}>
+        <div className={classes.center} style={style}>
             <CircularLoader />
         </div>
     )
+}
+
+LoadingMask.propTypes = {
+    style: PropTypes.object,
 }
 
 export default LoadingMask