+++ src/components/Item/VisualizationItem/Visualization/Visualization.js
@@ -1,17 +1,15 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
-import VisualizationPlugin from '@dhis2/data-visualizer-plugin'
 import i18n from '@dhis2/d2-i18n'
-import { D2Shim } from '@dhis2/app-runtime-adapter-d2'
 
-import DefaultPlugin from './DefaultPlugin'
+import LegacyPlugin from './LegacyPlugin'
 import MapPlugin from './MapPlugin'
-import LoadingMask from './LoadingMask'
+import DataVisualizerPlugin from './DataVisualizerPlugin'
 import NoVisualizationMessage from './NoVisualizationMessage'
 
-import getFilteredVisualization from './getFilteredVisualization'
-import getVisualizationConfig from './getVisualizationConfig'
+import getFilteredVisualization from '../../../../modules/getFilteredVisualization'
+import getVisualizationConfig from '../../../../modules/getVisualizationConfig'
 import {
     VISUALIZATION,
     MAP,
@@ -21,14 +19,9 @@ import {
 import { getVisualizationId } from '../../../../modules/item'
 import memoizeOne from '../../../../modules/memoizeOne'
 import { sGetVisualization } from '../../../../reducers/visualizations'
-import { UserSettingsCtx } from '../../../UserSettingsProvider'
 import { pluginIsAvailable } from './plugin'
 
 class Visualization extends React.Component {
-    state = {
-        pluginLoaded: false,
-    }
-
     constructor(props) {
         super(props)
 
@@ -38,18 +31,13 @@ class Visualization extends React.Component {
         this.memoizedGetVisualizationConfig = memoizeOne(getVisualizationConfig)
     }
 
-    onLoadingComplete = () => {
-        this.setState({ pluginLoaded: true })
-    }
-
     render() {
         const {
             visualization,
             activeType,
             item,
             itemFilters,
-            availableHeight,
-            availableWidth,
+            ...rest
         } = this.props
 
         if (!visualization) {
@@ -60,9 +48,9 @@ class Visualization extends React.Component {
             )
         }
 
-        const style = { height: availableHeight }
-        if (availableWidth) {
-            style.width = availableWidth
+        const style = { height: this.props.availableHeight }
+        if (this.props.availableWidth) {
+            style.width = this.props.availableWidth
         }
 
         const pluginProps = {
@@ -75,6 +63,7 @@ class Visualization extends React.Component {
                 item.type,
                 activeType
             ),
+            ...rest,
         }
 
         switch (activeType) {
@@ -82,32 +71,13 @@ class Visualization extends React.Component {
             case CHART:
             case REPORT_TABLE: {
                 return (
-                    <>
-                        {!this.state.pluginLoaded && (
-                            <div style={pluginProps.style}>
-                                <LoadingMask />
-                            </div>
+                    <DataVisualizerPlugin
+                        visualization={this.memoizedGetFilteredVisualization(
+                            pluginProps.visualization,
+                            pluginProps.itemFilters
                         )}
-                        <D2Shim d2Config={{}}>
-                            {({ d2 }) => (
-                                <VisualizationPlugin
-                                    d2={d2}
-                                    visualization={this.memoizedGetFilteredVisualization(
-                                        pluginProps.visualization,
-                                        pluginProps.itemFilters
-                                    )}
-                                    onLoadingComplete={this.onLoadingComplete}
-                                    forDashboard={true}
-                                    style={pluginProps.style}
-                                    userSettings={{
-                                        displayProperty: this.context
-                                            .userSettings
-                                            .keyAnalysisDisplayProperty,
-                                    }}
-                                />
-                            )}
-                        </D2Shim>
-                    </>
+                        style={pluginProps.style}
+                    />
                 )
             }
             case MAP: {
@@ -127,7 +97,7 @@ class Visualization extends React.Component {
                 return pluginIsAvailable(
                     pluginProps.activeType || pluginProps.item.type
                 ) ? (
-                    <DefaultPlugin {...pluginProps} />
+                    <LegacyPlugin {...pluginProps} />
                 ) : (
                     <NoVisualizationMessage
                         message={i18n.t(
@@ -139,8 +109,6 @@ class Visualization extends React.Component {
         }
     }
 }
-
-Visualization.contextType = UserSettingsCtx
 
 Visualization.propTypes = {
     activeType: PropTypes.string,