+++ src/components/Item/VisualizationItem/Visualization/plugin.js
@@ -7,28 +7,73 @@ import {
     itemTypeMap,
 } from '../../../../modules/itemTypes'
 import { getVisualizationId } from '../../../../modules/item'
-import getGridItemDomId from '../../../../modules/getGridItemDomId'
+import getVisualizationContainerDomId from '../../../../modules/getVisualizationContainerDomId'
+import { loadExternalScript } from '../../../../modules/loadExternalScript'
 
 //external plugins
-const itemTypeToExternalPlugin = {
+const itemTypeToGlobalVariable = {
     [MAP]: 'mapPlugin',
     [EVENT_REPORT]: 'eventReportPlugin',
     [EVENT_CHART]: 'eventChartPlugin',
 }
+
+const itemTypeToScriptPath = {
+    [MAP]: '/dhis-web-maps/map.js',
+    [EVENT_REPORT]: '/dhis-web-event-reports/eventreport.js',
+    [EVENT_CHART]: '/dhis-web-event-visualizer/eventchart.js',
+}
+
 const hasIntegratedPlugin = type => [CHART, REPORT_TABLE].includes(type)
 
-const getPlugin = type => {
+const getPlugin = async type => {
     if (hasIntegratedPlugin(type)) {
         return true
     }
-    const pluginName = itemTypeToExternalPlugin[type]
+    const pluginName = itemTypeToGlobalVariable[type]
 
-    return global[pluginName]
+    return await global[pluginName]
 }
 
-export const pluginIsAvailable = type => !!getPlugin(type)
+const fetchPlugin = async (type, baseUrl) => {
+    const globalName = itemTypeToGlobalVariable[type]
+    if (global[globalName]) {
+        return global[globalName] // Will be a promise if fetch is in progress
+    }
 
-export const loadPlugin = (plugin, config, credentials) => {
+    const scripts = []
+
+    if (type === EVENT_REPORT || type === EVENT_CHART) {
+        if (process.env.NODE_ENV === 'production') {
+            scripts.push('./vendor/babel-polyfill-6.26.0.min.js')
+            scripts.push('./vendor/jquery-3.3.1.min.js')
+            scripts.push('./vendor/jquery-migrate-3.0.1.min.js')
+        } else {
+            scripts.push('./vendor/babel-polyfill-6.26.0.js')
+            scripts.push('./vendor/jquery-3.3.1.js')
+            scripts.push('./vendor/jquery-migrate-3.0.1.js')
+        }
+    }
+
+    scripts.push(baseUrl + itemTypeToScriptPath[type])
+
+    const scriptsPromise = Promise.all(scripts.map(loadExternalScript)).then(
+        () => global[globalName] // At this point, has been replaced with the real thing
+    )
+
+    global[globalName] = scriptsPromise
+    return await scriptsPromise
+}
+
+export const pluginIsAvailable = type =>
+    hasIntegratedPlugin(type) || itemTypeToGlobalVariable[type]
+
+const loadPlugin = async (type, config, credentials) => {
+    if (!pluginIsAvailable(type)) {
+        return
+    }
+
+    const plugin = await fetchPlugin(type, credentials.baseUrl)
+
     if (!(plugin && plugin.load)) {
         return
     }
@@ -56,27 +101,24 @@ export const load = async (
     const config = {
         ...visualization,
         ...options,
-        el: getGridItemDomId(item.id),
+        el: getVisualizationContainerDomId(item.id),
     }
 
     const type = activeType || item.type
-    const plugin = getPlugin(type)
-
-    loadPlugin(plugin, config, credentials)
+    await loadPlugin(type, config, credentials)
 }
 
-export const resize = (item, isFullscreen) => {
-    const plugin = getPlugin(item.type)
-
-    if (plugin && plugin.resize) {
-        plugin.resize(getGridItemDomId(item.id), isFullscreen)
+export const resize = async (id, type, isFullscreen = false) => {
+    const plugin = await getPlugin(type)
+    if (plugin?.resize) {
+        plugin.resize(getVisualizationContainerDomId(id), isFullscreen)
     }
 }
 
-export const unmount = (item, activeType) => {
-    const plugin = getPlugin(activeType)
+export const unmount = async (item, activeType) => {
+    const plugin = await getPlugin(activeType)
 
     if (plugin && plugin.unmount) {
-        plugin.unmount(getGridItemDomId(item.id))
+        plugin.unmount(getVisualizationContainerDomId(item.id))
     }
 }